Whole menu option
The manu() function now accepts a boolean parameter. If the parameter
is set to true, then the whole menu tree is generated. If the parameter
is set to false, only the levels up to the current one are expandend.

This could be useful for websites which use dropdown menus using CSS.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>